refactor [#31] Sidebar와 NavigationBar 파일들 분류, import 구문 수정
