+++ src/components/gnb/index.ts
@@ -1,1 +1,2 @@
-export { default as Sidebar } from "./Sidebar";
+export { default as NavigationBar } from "./navigation-bar/NavigationBar";
+export { default as Sidebar } from "./sidebar/Sidebar";
