+++ src/components/gnb/navigation-bar/NavigationBarProfile.tsx
@@ -1,4 +1,4 @@
-import Icon from "../icon";
+import Icon from "@/components/icon";
 
 export default function NavigationBarProfile() {
   return (
