+++ src/components/gnb/navigation-bar/index.ts
@@ -1,0 +1,3 @@
+export { default as Drawer } from "./Drawer";
+export { default as NavigationBar } from "./NavigationBar";
+export { default as NavigationBarProfile } from "./NavigationBarProfile";
