+++ src/components/gnb/sidebar/index.ts
@@ -1,0 +1,2 @@
+export { default as Sidebar } from "./Sidebar";
+export { default as SidebarProfile } from "./SidebarProfile";
